requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-dotenv==1.4.2
djangorestframework==3.11.0
git+https://github.com/marco-silva0000/django-rest-framework-xml
idna==2.8
pkg-resources==0.0.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
sqlparse==0.3.0
tqdm==4.41.1
urllib3==1.25.7
